+++ CHANGELOG.md
@@ -38,9 +38,10 @@
     - `--archive` - shortcut for `--download-archive on`
 - New command `odoo-helper postgres stat-connections`
 - New shortcut `odoo-helper pg` for `odoo-helper postgres`
-- New command `odoo-helper git` with git-related helpers.
 - New command `odoo-helper git changed-addons` that displays
   list of addons that have been changed between two specified git revisions
+- New command `odoo-helper ci check-versions-git`.
+  The goal of this command is to be sure that addon version number was updated.
 
 ### Fix